+++ app/services/export/slack/insights/payload_service.rb
@@ -62,6 +62,7 @@ module Export
           ::Insights::VisibleQuery
             .new(relation: @insightable.insights)
             .resolve(insightable: @insightable)
+            .actual
             .joins(:entity)
             .hashable_pluck(
               :comments_count,